Crate libafl_targets

libafl_targets contains runtime code, injected in the target itself during compilation.

Re-exports

pub use sancov_pcguard::*;

pub use coverage::*;

pub use value_profile::*;

pub use cmplog::*;

pub use forkserver::*;

Modules

cmplog

CmpLog logs and reports back values touched during fuzzing. The values will then be used in subsequent mutations.

coverage

Coverage maps as static mut array

drcov

DrCov support for LibAFL frida mode, writing basic-block trace files to be read by coverage analysis tools, such as Lighthouse, bncov, dragondance, etc.

forkserver

Forkserver logic into targets

sancov_pcguard

LLVM PcGuard runtime for LibAFL.

value_profile

Value profile support for LibAFL

Constants

ACCOUNTING_MAP_SIZE

The size of the accounting maps

CMPLOG_MAP_H

The height of the CmpLog map

CMPLOG_MAP_W

The width of the CmpLog map

CMP_MAP_SIZE

The size of the cmps map

EDGES_MAP_SIZE

The size of the edges map