Module libafl_targets::cmplog

CmpLog logs and reports back values touched during fuzzing. The values will then be used in subsequent mutations.

Re-exports

pub use libafl_cmplog_map_ptr as CMPLOG_MAP_PTR;

pub use libafl_cmplog_map as CMPLOG_MAP;

pub use libafl_cmplog_enabled as CMPLOG_ENABLED;

Structs

CmpLogHeader

The header for CmpLog hits.

CmpLogInstruction

The operands logged during CmpLog.

CmpLogMap

A struct containing the CmpLog metadata for a LibAFL run.

CmpLogObserver

A CmpObserver observer for CmpLog

CmpLogRoutine

The routine arguments logged during CmpLog.

Constants

CMPLOG_KIND_INS

CmpLog instruction kind

CMPLOG_KIND_RTN

CmpLog return kind

CMPLOG_MAP_RTN_H

The hight of a cmplog routine map

CMPLOG_MAP_SIZE

The CmpLog map size

CMPLOG_RTN_LEN

The size of a logged routine argument in bytes

Statics

libafl_cmplog_enabled

Value indicating if cmplog is enabled.

libafl_cmplog_map

The global CmpLog map for the current LibAFL run.

libafl_cmplog_map_ptr

Pointer to the CmpLog map

Functions

__libafl_targets_cmplog_instructions⚠

Logs an instruction for feedback during fuzzing

Unions

CmpLogVals

Union of cmplog operands and routines